+++ NEWS.md
@@ -1,3 +1,20 @@
+# CRE 0.2.1 (2023-3-17)
+
+## Changed
+* Replace BATE with ATE in CATE Linear Decomposition.
+* Update `plot()` function (remove ATE, old BATE, and explicit AATEs).
+
+## Added
+* Code of Conduct.
+
+## Removed
+* Causal Tree benchmark in functional tests.
+
+## Bug fixes
+* Rank-Deficient Rule Matrix Issue (redundant rules).
+* Intervention Variables Filtering (ordered filtering).
+
+
 # CRE 0.2.0 (2023-1-19)
 
 ## Changed